Added a command for resetting all changes done to the local repositories
Until now we missing a command to reset all local repositories, created
by the checkout command, to the initial state after cloning them. To
undo changes done by the commands of the release tooling, it was needed
to delete all local repositories and to clone them again.

The new reset command offers now a faster method to return to the
initial state after cloning the repositories.